requirements.txt
absl-py==2.1.0
accelerate==1.3.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
anthropic==0.40.0
anyio==4.6.2.post1
appdirs==1.4.4
attrs==24.2.0
biopython==1.85
bitsandbytes==0.45.0
black==24.10.0
blessed==1.20.0
blis==1.1.0
Bottleneck
Brotli
catalogue==2.0.10
certifi
charset-normalizer
click==8.1.7
cloudpathlib==0.20.0
cloudpickle==3.1.1
cmake==3.31.4
cohere==5.12.0
colorlog==6.9.0
comm
confection==0.1.5
contourpy
cycler
cymem==2.0.10
datasets==3.1.0
debugpy
decorator
diffusers
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
evaluate==0.4.3
exceptiongroup
executing
fastapi==0.115.6
fastavro==1.9.7
ffmpy==0.4.0
filelock
fire==0.7.0
fonttools
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
gradio==5.8.0
gradio_client==1.5.1
greenlet==3.1.1
grpcio==1.68.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.0
httpx-sse==0.4.0
huggingface_hub
idna
interegular==0.3.3
ipdb==0.13.13
ipykernel
ipython==8.30.0
jedi
Jinja2
jiter==0.8.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
llvmlite==0.44.0
lm-format-enforcer==0.9.8
loralib==0.1.2
Mako==1.3.7
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.9.2
matplotlib-inline
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft
mkl_random
mpmath
msgpack
multidict
multiprocess
murmurhash
mypy-extensions
nest_asyncio
ninja==1.11.1.3
nltk==3.9.1
numba==0.61.0
numexpr
numpy
nvidia-ml-py==12.560.30
openai==1.55.3
optuna==4.1.0
orjson==3.10.12
outlines==0.0.34
packaging
pandas
parameterized==0.9.0
parso
pathspec==0.12.1
peft==0.14.0
pexpect
pickleshare
pillow
platformdirs
plotly==5.24.1
ply==3.11
preshed==3.0.9
progress==1.6
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.1
py-cpuinfo==9.0.0
pyarrow==18.1.0
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
Pygments==2.18.0
pyparsing
PySocks
python-dateutil
python-dotenv==1.0.1
python-multipart==0.0.19
pytz
PyYAML
pyzmq
ray==2.41.0
referencing==0.36.2
regex
requests
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.2
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.5.2
scipy
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
shellingham==1.5.4
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.5.0
starlette==0.41.3
sympy
tenacity
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thinc==8.3.3
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenize_rt==6.1.0
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.2.1
tqdm
transformers==4.46.3
triton==2.2.0
trl==0.13.0
typer==0.15.1
types-requests==2.32.0.20241016
uvicorn==0.32.1
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.0
xformers==0.0.25
xxhash==3.5.0
yarl==1.18.3
langchain==0.2.17
langchain_openai==0.2.17
